feat: 타입을 도메인 별로 파일 분리
