+++ src/types/cartItem.type.ts
@@ -1,0 +1,9 @@
+export type CartItem = {
+  id: number;
+  quantity: number;
+  product: {
+    id: number;
+    name: string;
+    price: number;
+  };
+};
