+++ src/types/product.type.ts
@@ -1,0 +1,9 @@
+import { Category } from './filter.type';
+
+export type Product = {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  category: Category;
+};
